fixes age calculation bug, turns out the createdOn date only moves if the change is edited (not rebased) so instead I'm tracking when jenkins last ran CI on the test to indicate a rebase has happened.
The age calculation isn't straight forward. I have decided to look for
a timestamp in the patchset which indicates that someone had "done
work" on the patchset (beyond commenting) the only reliable way I could
think was to look for the Jenkins CI comment. The createdOn date might
not change if the update was only a rebase.